+++ apps/extension/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [2.0.0] - 09-07-2025
 
+- 🐞 fix: translation for `zh` language
 - 🔧 chore: add tests for utils
 - ✨ feat: detect and display private IPs (v4/v6) using internal-ip
 - ✨ feat: detect and display public IP info with geo location, timezone, connection and country flag